third_party/blink/web_tests/virtual/document-isolation-policy/external/wpt/html/document-isolation-policy/credentialless-cache-storage.https.tentative.window_dedicated_worker-expected.txt

This is a testharness.js-based test.
[FAIL] [dedicated_worker] none => isolate-and-credentialless
  assert_equals: expected "error" but got "retrieved"
[FAIL] [dedicated_worker] none => isolate-and-require-corp
  assert_equals: expected "error" but got "retrieved"
[FAIL] [dedicated_worker] none => isolate-and-require-corp (omit)
  assert_equals: expected "error" but got "retrieved"
[FAIL] [dedicated_worker] isolate-and-credentialless => isolate-and-require-corp
  assert_equals: expected "error" but got "retrieved"
Harness: the test ran to completion.
